pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcriptor"
version = "0.1.0"
description = "Call center audio transcription and diarization system"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.1",
    "watchdog>=4.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydub>=0.25",
    "click>=8.1",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=24.0",
    "ruff>=0.3",
]

[project.scripts]
transcriptor = "transcriptor.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
